Create APITraining
An Event Apart the Web app that we’re about to go through now. It’s called Front Row. 
The idea of the Web app is to take a movie title inputted by the user, run it through TMDb’s API, 
and return the relevant poster. The user could then share it or save it to their computer.